requirements.txt
# Open Medical Secretary - Dependencies
# Install with: pip install -r requirements.txt

# Core framework
pipecat-ai[silero,whisper,daily]==0.0.98

# Web interface
flask>=3.0.0

# Audio processing
pyaudio>=0.2.13
scipy>=1.11.0
numpy>=1.24.0

# AI Services
ollama>=0.3.0
TTS>=0.22.0

# Telephony (optional)
websockets>=12.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
